requirements.txt
torch
torchvision
fastapi[standard]
pydantic
uvicorn[standard]
google-genai
tinydb
